auth/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Sarim95khan <sarim95khan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlmodel = "^0.0.19"
uvicorn = "^0.30.1"
alembic = "^1.13.2"
psycopg = {extras = ["binary"], version = "^3.2.1"}
passlib = "^1.7.4"
python-jose = "^3.3.0"
rsa = "^4.9"
aiokafka = "^0.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
